Finish 5-1
